+++ tests/Neo.L2.ForcedInclusion.UnitTests/Usings.cs
@@ -1,3 +1,4 @@
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using Neo.L2;
 global using Neo.L2.ForcedInclusion;
+global using Neo.L2.Telemetry;